+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fix only SHA1 checksum is accepted when uploading resources [#2808](https://github.com/opendatateam/udata/pull/2808)
 - Fix organization metrics count [#2811](https://github.com/opendatateam/udata/pull/2811)
 - Fix setuptools version used in CI [#2813](https://github.com/opendatateam/udata/pull/2813)
+- Add `udata harvest clean` command [#2812](https://github.com/opendatateam/udata/pull/2812)
 
 ## 6.0.1 (2023-01-18)